requirements.txt
motor~=3.7.1
pydantic[email]~=2.12.3
pydantic-settings~=2.11.0
grpcio~=1.76.0
grpcio-tools~=1.76.0
aiohttp~=3.13.2
protobuf~=6.33.1
python-jose~=3.5.0
passlib~=1.7.4
fastapi~=0.115.0
uvicorn[standard]~=0.30.1